demo: split dialog demo into window dialog and overlay dialog pages
- Remove DialogDemo page and DialogDemoViewModel
- Add WindowDialogDemo page with Default/Custom Window Dialog tabs
- Add OverlayDialogDemo page with Default/Custom Overlay Dialog tabs
- Move sub-view models to their respective new ViewModel files
- Update menu to show 'Window Dialog' and 'Overlay Dialog' entries
- Add MenuKeyWindowDialog and MenuKeyOverlayDialog constants
- Update navigation routing in MainViewViewModel
