+++ demo/Ursa.Demo/Pages/OverlayDialogDemo.axaml.cs
@@ -2,9 +2,9 @@ using Avalonia.Controls;
 
 namespace Ursa.Demo.Pages;
 
-public partial class DialogDemo : UserControl
+public partial class OverlayDialogDemo : UserControl
 {
-    public DialogDemo()
+    public OverlayDialogDemo()
     {
         InitializeComponent();
     }
