+++ demo/Ursa.Demo/Pages/WindowDialogDemo.axaml.cs
@@ -1,0 +1,11 @@
+using Avalonia.Controls;
+
+namespace Ursa.Demo.Pages;
+
+public partial class WindowDialogDemo : UserControl
+{
+    public WindowDialogDemo()
+    {
+        InitializeComponent();
+    }
+}
