+++ demo/Ursa.Demo/ViewModels/Pages/MainViewViewModel.cs
@@ -48,7 +48,8 @@ public partial class MainViewViewModel : ViewModelBase
             MenuKeys.MenuKeyDateOffsetRangePicker => new DateOffsetRangePickerDemoViewModel(),
             MenuKeys.MenuKeyDateTimeOffsetPicker => new DateTimeOffsetPickerDemoViewModel(),
             MenuKeys.MenuKeyDescriptions => new DescriptionsDemoViewModel(),
-            MenuKeys.MenuKeyDialog => new DialogDemoViewModel(),
+            MenuKeys.MenuKeyWindowDialog => new WindowDialogDemoViewModel(),
+            MenuKeys.MenuKeyOverlayDialog => new OverlayDialogDemoViewModel(),
             MenuKeys.MenuKeyDisableContainer => new DisableContainerDemoViewModel(),
             MenuKeys.MenuKeyDivider => new DividerDemoViewModel(),
             MenuKeys.MenuKeyDrawer => new DrawerDemoViewModel(),
